fix: renamed and linted
