+++ src/components/404/404.scss
@@ -1,13 +1,8 @@
-
 .error {
     width: 100%;
     height: 100vh;
-    background-image: url("../../assets/images/Error-404.jpg");
-    /* Center and scale the image nicely */
+    background-image: url('../../assets/images/Error-404.jpg');
     background-position: center;
     background-repeat: no-repeat;
     background-size: cover;
-
-    
-
 }
