+++ src/pages/Home/home.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 
 import Title from '../../components/Title/title';
 import Navbar from '../../components/NavBar/navbar';
-import About from '../../components/About Us/about';
+import About from '../../components/AboutUs/AboutUs';
 import PastWinners from '../../components/Past-Winners/past-winners';
 import TeamPics from '../../components/TeamPics/teampics';
 import Footer from '../../components/Footer/footer';
